Sepperate the bsp specific functions into their own files.
This will hopefully make it easier to switch to different parallel / communication methods.
